Cleanup route and setup code
