--- src/apiRouter.ts
@@ -1,10 +1,0 @@
-import AsyncRouter from './AsyncRouter';
-import { Request, Response } from 'express';
-
-const api = new AsyncRouter();
-
-api.get('/', async (req: Request, res: Response) => {
-  res.json({ works: true });
-});
-
-export default api;
