+++ src/config/index.ts
@@ -1,11 +1,11 @@
 export const isProduction = process.env.NODE_ENV === 'production';
 export const isDevelopment = !isProduction;
-
 export const PORT = envOrDefault('PORT', '1337');
-export const DB_CONNECTION = envOrDefault(
+export const DB_CONNECTION_URL = envOrDefault(
   'DB_CONNECTION',
   'postgresql://dev_user:dev_password@localhost/dev_db'
 );
+export const API_PREFIX = envOrDefault('API_PREFIX', '/api');
 
 function envOrDefault(
   key: string,
