+++ src/database/database.ts
@@ -1,10 +1,10 @@
 import knex from 'knex';
 
-import { DB_CONNECTION } from '../config';
+import { DB_CONNECTION_URL } from '../config';
 
 const database = knex({
   client: 'pg',
-  connection: DB_CONNECTION,
+  connection: DB_CONNECTION_URL,
 });
 
 export function migrateLatest(): PromiseLike<any> {
