+++ src/loaders/database.ts
@@ -1,0 +1,5 @@
+import { migrateLatest } from '../database';
+
+export async function migrateDatabase() {
+  await migrateLatest();
+}
